Improve ~more~ comment viewing operation
