--- ui/component/expandable/index.js
@@ -1,3 +1,0 @@
-import Expandable from './view';
-
-export default Expandable;
